Functions on the driver should be protected
There is no need for many of these methods to be private, and in fact it
makes it impossible to extend the driver sensibly if they are.

The visibility features of PHP mean that any function called in a parent
class by a child class where that parent function calls a private method
means that it is impossible to override the private methods.

This means that it is not possible, for example, to apply bug fixes
locally.